shared_service/requirements.txt
Flask
PyJWT
Flask-SQLAlchemy
SQLAlchemy
Flask-Migrate
python-dotenv